remove the scenario from the library
